@@ -1,5 +1,5 @@
 module golibs
 
-go 1.14
+go 1.22
 
 require golang.org/x/text v0.3.2
